+++ release_notes.md
@@ -1,4 +1,7 @@
-# :mega: borb release 2.0.31
+# :mega: borb release 2.0.32
 
 This release is a small bugfix release:
-- include `'` as a valid postfix operator (improve text reading by 2%)
+- fixed `Checkbox` to produce valid PDF syntax
+- refactored a few tests to have consistent naming
+- included `ProgressBar` and `ProgressSquare` classes to make it easier to make reports
+- included `TableUtil` and `TableUtil.from2DArray` to make it easier to build tables